requirements.txt
asttokens==2.4.1
azure-common==1.1.28
azure-core==1.30.1
azure-identity==1.16.0
azure-search-documents==11.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
comm==0.2.2
cryptography==42.0.7
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.1
executing==2.0.1
idna==3.7
ipykernel==6.29.4
ipython==8.24.0
isodate==0.6.1
jedi==0.19.1
jupyter_client==8.6.1
jupyter_core==5.7.2
matplotlib-inline==0.1.7
msal==1.28.0
msal-extensions==1.1.0
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.1
portalocker==2.8.2
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==26.0.3
regex==2024.5.10
requests==2.31.0
six==1.16.0
stack-data==0.6.3
tenacity==8.3.0
tiktoken==0.6.0
tornado==6.4
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
